+++ HISTORY.md
@@ -8,6 +8,9 @@
 * **Notable Fixes:**
     * made scrollbars in the wardrobe and jukebox more responsive
 
+### 2021-04-22
+* store, gui: first pass at deserializing multi-scene projects
+
 ### 2021-04-21
 * store, gui: refactored project loading structure